•TimeManager.cs is working now. -TimeManager.cs has Pre countdown and main game countdown in it. -TimerManager.cs changed to TimeManager.cs.
•UIManager.cs is working now.
UIManager.cs is managing the UI Elements for now. For Change text and set them visible or not.

•AIController.cs is working now.
AIController.cs is picking randomly one tile for blast. Working totally randomly.

•ScoreManager.cs is working now.
ScoreManager.cs is managing the Player score and AI score and at the end of the game checking for the Winner.

•GridController.cs has some changes.
For AI setting added on GridController.cs so we can split AI GridController.cs.

•Scene Changes
-UI Elements for Pre time, game time, scores and Winner added.
-One more GridController.cs added on scene for AI